requirements.txt
# CATKC-Net Requirements
# Install with: pip install -r requirements.txt

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Image Processing
opencv-python>=4.7.0
scikit-image>=0.20.0
Pillow>=9.5.0

# Numerical
numpy>=1.24.0
scipy>=1.10.0

# Training Utilities
tqdm>=4.65.0
tensorboard>=2.13.0

# Architecture Utilities
einops>=0.6.1

# Perceptual Loss
lpips>=0.1.4

# Visualization
matplotlib>=3.7.0

# Config
pyyaml>=6.0

# Optional: for notebook use
# jupyter
# ipywidgets